try and get icons working
