+++ source/menu-entry.c
@@ -1,6 +1,8 @@
 #include "common.h"
 #include "parsing/shortcut.h"
 
+#include <libconfig.h>
+
 void menuEntryInit(menuEntry_s* me, MenuEntryType type)
 {
 	memset(me, 0, sizeof(*me));
@@ -47,18 +49,18 @@ char* normalizePath(const char* path) {
 	return copy;
 }
 
-static bool menuEntryLoadExternalIcon(menuEntry_s* me, const char* filepath) {
-	if (true)
-		return false;
-
+bool menuEntryLoadExternalIcon(menuEntry_s* me, const char* filepath) {
 	struct stat fileStat;
 
+	LOG("Loading Icon: %s", filepath);
 	if (stat(filepath, &fileStat) == -1)
 		return false;
 
-	FILE* icon = fopen(filepath, "rb");
-	if (!icon)
+	FILE* iconFile = fopen(filepath, "rb");
+	if (!iconFile) {
+		fclose(iconFile);
 		return false;
+	}
 
 	if (!me->icon)
 	{
@@ -66,20 +68,24 @@ static bool menuEntryLoadExternalIcon(menuEntry_s* me, const char* filepath) {
 		C3D_TexInit(me->icon, 64, 64, GPU_RGB565);
 	}
 
-	int fileNo = fileno(icon);
-	Tex3DS_Texture tex = Tex3DS_TextureImportFD(fileNo, me->icon, NULL, true);
+	Tex3DS_Texture texture = Tex3DS_TextureImportStdio(iconFile, me->icon, NULL, false);
+	fclose(iconFile);
 
-	// Copy 48x48 -> 64x64
-	u16* dest = (u16*)me->icon->data + (64-48)*64;
-	// u16* src = (u16*)data;
+	if (!texture)
+		return false;
 
-	int j;
-	for (j = 0; j < 48; j += 8)
-	{
-		// memcpy(dest, src, 48*8*sizeof(u16));
-		// src += 48*8;
-		dest += 64*8;
-	}
+	const Tex3DS_SubTexture* subTexture = Tex3DS_GetSubTexture(texture, 0);
+
+	if (subTexture->width != 64)
+		return false;
+
+	if (subTexture->height != 64)
+		return false;
+
+	// Delete the t3x object since we don't need it
+	Tex3DS_TextureFree(texture);
+
+	return true;
 }
 
 static bool menuEntryLoadEmbeddedSmdh(menuEntry_s* me)
@@ -127,6 +133,176 @@ static void fixSpaceNewLine(char* buf)
 	} while (lastc);
 }
 
+void menuEntryFileAssocLoad(const char* filepath)
+{
+	bool success = false;
+	bool iconLoaded = false;
+
+	menuEntry_s* entry = NULL;
+
+	config_setting_t* fileAssoc = NULL;
+	config_setting_t* targets = NULL;
+	config_setting_t* target = NULL;
+	config_setting_t* appArguments = NULL;
+	config_setting_t* targetArguments = NULL;
+
+	config_t config = {0};
+
+	int targetsLength = 0;
+	int argsLength = 0;
+	int index = 0;
+
+	char appPath[PATH_MAX + 8];
+	char mainIconPath[PATH_MAX + 1];
+	char targetIconPath[PATH_MAX + 1];
+	char targetFileExtension[PATH_MAX + 1];
+	char targetFilename[PATH_MAX + 1];
+
+	char appAuthor[ENTRY_AUTHORLENGTH + 2];
+	char appDescription[ENTRY_DESCLENGTH + 2];
+
+	const char* stringValue = NULL;
+
+	config_init(&config);
+
+	memset(appPath, 0, sizeof(appPath));
+	memset(mainIconPath, 0, sizeof(mainIconPath));
+	memset(appAuthor, 0, sizeof(appAuthor));
+	memset(appDescription, 0, sizeof(appDescription));
+
+
+	if (!fileExists(filepath))
+		return;
+
+	if (config_read_file(&config, filepath)) {
+
+		fileAssoc = config_lookup(&config, "fileassoc");
+
+		if (fileAssoc != NULL) {
+			if (config_setting_lookup_string(fileAssoc, "app_path", &stringValue))
+				snprintf(appPath, sizeof(appPath) - 1, "%s%s", menuGetRootBasePath(), stringValue);
+
+			if (config_setting_lookup_string(fileAssoc, "icon_path", &stringValue))
+                snprintf(mainIconPath, sizeof(mainIconPath) - 1, "%s%s", menuGetRootBasePath(), stringValue);
+			LOG("Main Icon: %s", mainIconPath == NULL ? "None" : mainIconPath);
+			appArguments = config_setting_lookup(fileAssoc, "app_args");
+            targets = config_setting_lookup(fileAssoc, "targets");
+
+			if (appPath[0] && targets) {
+				targetsLength = config_setting_length(targets);
+
+				if (targetsLength > 0) {
+					entry = menuCreateEntry(ENTRY_TYPE_FILE);
+					success = false;
+
+					if (entry) {
+						strncpy(entry->path, appPath, sizeof(entry->path) - 1);
+						entry->path[sizeof(entry->path) - 1] = 0;
+
+						stringValue = getSlash(appPath);
+
+						if (stringValue[0] == '/')
+							stringValue++;
+
+						if (menuEntryLoad(entry, stringValue, false)) {
+							strncpy(appAuthor, entry->author, sizeof(appAuthor));
+							appAuthor[sizeof(appAuthor) - 1] = 0;
+
+							strncpy(appDescription, entry->description, sizeof(appDescription));
+							appDescription[sizeof(appDescription) - 1] = 0;
+
+							success = true;
+						}
+
+						menuDeleteEntry(entry);
+						entry = NULL;
+					}
+
+					if (success) {
+						for (index = 0; index < targetsLength; index++) {
+							target = config_setting_get_elem(targets, index);
+
+							if (target == NULL)
+								continue;
+
+							memset(targetIconPath, 0, sizeof(targetIconPath));
+							memset(targetFileExtension, 0, sizeof(targetFileExtension));
+							memset(targetFilename, 0, sizeof(targetFilename));
+
+							if (config_setting_lookup_string(target, "icon_path", &stringValue))
+								snprintf(targetIconPath, sizeof(targetIconPath) - 1, "%s%s", menuGetRootBasePath(), stringValue);
+
+							if (config_setting_lookup_string(target, "file_extension", &stringValue))
+								strncpy(targetFileExtension, stringValue, sizeof(targetFileExtension) - 1);
+
+							if (config_setting_lookup_string(target, "filename", &stringValue))
+								strncpy(targetFilename, stringValue, sizeof(targetFilename) - 1);
+
+							targetArguments = config_setting_lookup(target, "app_args");
+
+							if ((targetFileExtension[0] != 0) == (targetFilename[0] != 0))
+								continue;
+
+							entry = menuCreateEntry(ENTRY_TYPE_FILEASSOC);
+							iconLoaded = false;
+
+							if (entry) {
+								strncpy(entry->path, appPath, sizeof(entry->path));
+								entry->path[sizeof(entry->path) - 1] = 0;
+
+								strncpy(entry->author, appAuthor, sizeof(entry->author));
+								entry->author[sizeof(entry->author) - 1] = 0;
+
+								strncpy(entry->description, appDescription, sizeof(entry->description));
+								entry->description[sizeof(entry->description) - 1] = 0;
+
+								if (targetFileExtension[0]) {
+									entry->fileAssocType = 0;
+									strncpy(entry->fileAssocStr, targetFileExtension, sizeof(entry->fileAssocStr));
+								} else if (targetFilename[0]) {
+									entry->fileAssocType = 1;
+									strncpy(entry->fileAssocStr, targetFilename, sizeof(entry->fileAssocStr));
+								}
+
+								entry->fileAssocStr[sizeof(entry->fileAssocStr) - 1] = 0;
+
+								if (targetIconPath[0])
+									iconLoaded = menuEntryLoadExternalIcon(entry, targetIconPath);
+
+								if (!iconLoaded && mainIconPath[0])
+									iconLoaded = menuEntryLoadExternalIcon(entry, mainIconPath);
+
+								argData_s* argData = &entry->args;
+								argData->dst = (char*)&argData->buf[1];
+
+								launchAddArg(argData, entry->path);
+
+								config_setting_t* configArgs = targetArguments ? targetArguments : appArguments;
+								if (configArgs) {
+									argsLength = config_setting_length(configArgs);
+									for (int argument = 0; argument < argsLength; argument++) {
+										stringValue = config_setting_get_string_elem(configArgs, argument);
+
+										if (stringValue == NULL)
+											continue;
+
+										launchAddArg(argData, stringValue);
+									}
+								}
+							}
+
+							if (entry)
+								menuFileAssocAddEntry(entry);
+						}
+					}
+				}
+			}
+		}
+	}
+
+	config_destroy(&config);
+}
+
 bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 {
 	static char tempbuf[PATH_MAX+1];
@@ -154,7 +330,7 @@ bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 			snprintf(tempbuf, sizeof(tempbuf)-1, "%.*s/%.*s.3dsx", sizeof(tempbuf)/2, me->path, sizeof(tempbuf)/2-7, name);
 			found = fileExists(tempbuf);
 		}
-		LOG("Found: %d, File Assoc Entries %d", found, menuFileAssoc->nEntries);
+
 		if (!found && menuFileAssoc->nEntries > 0) {
 			fileAssocFlag = true;
 
@@ -175,7 +351,6 @@ bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 							continue;
 
 						snprintf(tempbuf, sizeof(tempbuf) - 1, "%.*s/%.*s", (int)sizeof(tempbuf) / 2, me->path, (int)sizeof(tempbuf) / 2 - 7, dp->d_name);
-						LOG("File Association: %s", tempbuf);
 						found = fileExists(tempbuf);
 
 						if (found)
@@ -209,13 +384,11 @@ bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 		if (shortcut)
 		{
 			if (R_FAILED(shortcutCreate(&sc, me->path))) {
-				LOG("Failed to create shortcut!");
 				return false;
 			}
 
 			if (!fileExists(sc.executable))
 			{
-				LOG("Shortcut executable does not exist.");
 				shortcutFree(&sc);
 				return false;
 			}
@@ -339,8 +512,9 @@ bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 			strptr = getExtension(tempbuf);
 			strncpy(strptr, ".t3x", sizeof(tempbuf) - 1 - ((ptrdiff_t)strptr - (ptrdiff_t)tempbuf));
 
-			bool iconLoaded = false;
-			/* TODO: load an icon */
+			bool iconLoaded = menuEntryLoadExternalIcon(me, tempbuf);
+			if (!iconLoaded && !fileAssocEntry->icon)
+				LOG("I don't know?");
 
 			/* attempt to load the smdh from entry->path with extension .smdh */
 			/* on failure, use the config from the entry */
@@ -374,7 +548,7 @@ bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut)
 			for (u32 argIndex = 0; argIndex < argDataAssoc->buf[0]; argIndex++, argSource += strlen(argSource) + 1) {
 				if (argIndex) {
 					strptr = strchr(argSource, '%');
-					if (strptr && strptr[0] && strptr[1] && (strptr == argSource || strptr[-1] != "\\")) {
+					if (strptr && strptr[0] && strptr[1] && (strptr == argSource || strptr[-1] != '\\')) {
 						if (strptr[1] == 'f') {
 							memset(tempbuf, 0, sizeof(tempbuf));
 							snprintf(tempbuf, sizeof(tempbuf) - 1, "%.*s%s%s", (int)((uintptr_t)strptr - (uintptr_t)argSource), argSource, me->path, &strptr[2]);
