+++ source/menu.h
@@ -39,8 +39,8 @@ struct menuEntry_s_tag
 	char description[ENTRY_DESCLENGTH+1];
 	char author[ENTRY_AUTHORLENGTH+1];
 
-    bool fileAssocType; 			//< 0 file_extension, 1 = filename
-    char fileAssocStr[PATH_MAX+1];	//< file_extension/filename
+    bool fileAssocType; 					//< 0 file_extension, 1 = filename
+    char fileAssocStr[PATH_MAX + 1];		//< file_extension/filename
 
 	smdh_s smdh;
 	descriptor_s descriptor;
@@ -55,11 +55,15 @@ struct menuEntry_s_tag
 	bool isStarred;
 };
 
+menuEntry_s* menuCreateEntry(MenuEntryType type);
+void menuDeleteEntry(menuEntry_s* me);
+
 void menuEntryInit(menuEntry_s* me, MenuEntryType type);
 void menuEntryFree(menuEntry_s* me);
 bool menuEntryLoad(menuEntry_s* me, const char* name, bool shortcut);
 void menuEntryParseSmdh(menuEntry_s* me);
 void menuEntryFileAssocLoad(const char* filepath);
+bool menuEntryLoadExternalIcon(menuEntry_s* me, const char* filepath);
 
 struct menu_s_tag
 {
