+++ source/ui/menu.h
@@ -6,5 +6,6 @@
 void menuUpdate(void);
 void menuDrawTop(float iod);
 void menuDrawBot(void);
+void menuLoadFileAssoc(void);
 
 float menuDrawEntry(menuEntry_s* me, float x, float y, bool selected);
